feat: usage的tid_list也进行once_cell避免反复判断
