+++ test_build.sh
@@ -3,4 +3,5 @@ cargo fmt
 sed -i 's/edition = "2024"/edition = "2021"/g' Cargo.toml
 rm -rf output
 rm -rf $(find ./target/aarch64-linux-android/debug -name "*thread-opt*")
-python3 ./make.py build --debug --nightly
+
+python3 ./make.py build --debug --nightly || rm -rf $(find ./target/aarch64-linux-android/debug -name "*mimalloc*")
